# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gigabee decentralized AI inference network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /credits/balance:
    get:
      operationId: getCreditsBalance
      tags: [credits]
      summary: Get user credit balance
      responses:
        "200":
          description: Credit balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreditsBalance"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /credits/transactions:
    get:
      operationId: getCreditTransactions
      tags: [credits]
      summary: Get credit transaction history
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreditTransaction"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chat/completions:
    post:
      operationId: createChatCompletion
      tags: [chat]
      summary: Create a chat completion (OpenAI-compatible)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
        "402":
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /referrals/code:
    get:
      operationId: getReferralCode
      tags: [referrals]
      summary: Get user's personal referral link/code
      responses:
        "200":
          description: Referral code and link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralCode"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /referrals/stats:
    get:
      operationId: getReferralStats
      tags: [referrals]
      summary: Get referral program statistics
      responses:
        "200":
          description: Referral stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralStats"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CreditsBalance:
      type: object
      required: [credits, usdValue]
      properties:
        credits:
          type: integer
        usdValue:
          type: number
...
    CreditTransaction:
      type: object
      required: [id, type, credits, description, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [topup, debit, refund]
        credits:
          type: integer
        description:
          type: string
        createdAt:
          type: string
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          enum: [bee-nano, bee-hover, bee-glide]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        stream:
          type: boolean
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
    ChatCompletionResponse:
      type: object
      required: [id, model, content, creditsDeducted, workerId, tokensUsed]
      properties:
        id:
          type: string
        model:
          type: string
        content:
          type: string
        creditsDeducted:
          type: integer
        workerId:
          type: string
        tokensUsed:
          type: integer
...
    ReferralCode:
      type: object
      required: [code, link]
      properties:
        code:
          type: string
        link:
          type: string
    ReferralStats:
      type: object
      required: [totalReferred, activeReferrals, earningsUsd, earningsCap]
      properties:
        totalReferred:
          type: integer
        activeReferrals:
          type: integer
        earningsUsd:
          type: number
        earningsCap:
          type: number
        referrals:
          type: array
          items:
            $ref: "#/components/schemas/ReferralEntry"
    ReferralEntry:
      type: object
      required: [id, joinedAt, spendUsd, yourEarningsUsd, active]
      properties:
        id:
          type: string
        joinedAt:
          type: string
        spendUsd:
          type: number
        yourEarningsUsd:
          type: number
        active:
          type: boolean
...